pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "glb2stl"
dynamic = ["version"]
description = "FastAPI service to convert GLB (non-Draco) to STL with unit & axis normalization."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Nayan Shah", email = "42316364+snayan06@users.noreply.github.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]

dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "trimesh>=4.7.4",
  "numpy>=2.3.2",
  "pydantic-settings>=2.10.1",
  "python-multipart>=0.0.20",
  "python-json-logger>=3.3.0",
  "scipy>=1.16.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "pytest-cov>=6.2.1",
  "httpx>=0.28.1",
  "ruff>=0.12.11",
  "mypy>=1.17.1",
  "pre-commit>=4.3.0",
]

[project.urls]
Documentation = "https://github.com/Nayan Shah/glb2stl#readme"
Issues        = "https://github.com/Nayan Shah/glb2stl/issues"
Source        = "https://github.com/Nayan Shah/glb2stl"

# Console script: `glb2stl-api`
[project.scripts]
glb2stl-api = "glb2stl.__main__:main"

# Version is read from a file (no Git needed)
[tool.hatch.version]
path = "src/glb2stl/__about__.py"

# src/ layout
[tool.hatch.build.targets.wheel]
packages = ["src/glb2stl"]

# ---- Dev tool configs ----
[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP","B","C4","SIM"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_unused_ignores = true
check_untyped_defs = true
pretty = true

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1"
testpaths = ["tests"]

[tool.coverage.run]
source_pkgs = ["glb2stl", "tests"]
branch = true
parallel = true
omit = ["src/glb2stl/__about__.py"]

[tool.coverage.paths]
glb2stl = ["src/glb2stl", "*/glb2stl/src/glb2stl"]
tests = ["tests", "*/glb2stl/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
